requirements.txt
flask==2.0.1
requests==2.28.2
python-dotenv==0.21.1
flask-cors==3.0.10
websocket-client==1.7.0


